chore: clean up clippy warnings (166 -> 50)
Applied `cargo clippy --fix` across workspace to resolve stylistic lints:
- 78 collapsible_if -> 14 remaining (nested `if let` chains requiring
  let_chains, stable in 1.88+ but MSRV is 1.85)
- 11 map_or simplifications (.map_or(false, _) -> .is_some_and(_))
- 11 derivable_impls (manual Default -> #[derive(Default)])
- 6 io::Error::new(Other, _) -> io::Error::other(_)
- 5 cloned_ref_to_slice_refs -> std::slice::from_ref
- 4 field_reassign_with_default
- 6 redundant_closure
- misc: redundant returns, unused imports, needless_borrows, etc.

Net change: -120 lines across 60 files. No behavior changes.

Remaining 50 warnings are judgment calls (too_many_arguments 8/7,
large_err_variant, complex_type) or require let_chains for clean fixes.

+++ crates/archon-tui/tests/pane_tests.rs
@@ -232,7 +232,7 @@ fn split_below_minimum_size_refused() {
         for r in &rects {
             // rects may be adjusted to enforce minimum
             assert!(
-                r.height >= PANE_MIN_ROWS as u16 || r.height == 0,
+                r.height >= PANE_MIN_ROWS || r.height == 0,
                 "pane height must be >= minimum: {}",
                 r.height
             );
